// src/constants/enums.ts
export enum buttonStyles {
    PRIMARY = "PRIMARY",
    SECONDARY = "SECONDARY",
    TERTIARY= "TERTIARY"
}
export enum backGroundColors {
    FACEBOOK = "#E7EAF4",
    GOOGLE = "#FAE9EA",
    APPLE = "#e3e3e3"
}

export enum foreGroundColors {
    FACEBOOK = "#4765A9",
    GOOGLE = "#DD4D44",
    APPLE = "#363636"
}